+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Status bar no longer displays an unwanted trailing character
 - Edit menu items are disabled unless the chat or image input text field has the keyboard focus
 - Improve speech error handling to fix crashes
+- No longer try to use AHI if the speech system doesn't use it
 
 ## 3.0.0 (2026-07-01)